[path_provider] Move Windows FFI behind a conditional import (#3056)
Moves the real implementation of path_provider_windows behind a
conditional export, instead exporting a stub on platforms that don't
support dart:ffi. This avoids build breakage in web projects that have
transitive dependencies on path_provider (and thus path_provider_windows
due to manual endorsement).

This will no longer be necessary once
https://github.com/flutter/flutter/issues/52267 is fixed, since only
Windows builds will ever need to have code-level dependency on
path_provider_windows.
